commit before handling audio table view nil error.
